Workout creation improvements
Saving set weights with floating point numbers now possible

 Input component is now controlled

Added delete cascade to workout, exercise and set models in DB schema

+++ components/pages/workout/exerciseInputBlock.tsx
@@ -51,7 +51,7 @@ const ExerciseInputBlock = ({ exerciseData }: Props) => {
     const changeSetWeight = (index: number, event: ChangeEvent<HTMLInputElement>) => {
         const { value } = event.target;
         const setInputs = [...sets];
-        setInputs[index].weight = parseInt(value);
+        setInputs[index].weight = parseFloat(value);
         setSets(setInputs);
     };
